refactor: reorganize tests and consolidate testing utilities
- Move test-only data fetching to tests/testing_utils.py
- Relocate geodesic functions to models/geo.py and remove utils/
- Organize tests into unit/ and integration/ directories
- Update all imports and benchmarks to reflect the new structure
- Maintain identical test count (118) and bit-wise parity
