+++ pyproject.toml
@@ -122,10 +122,10 @@ warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
 disallow_untyped_calls = false
-exclude = ["worktrees", ".venv", "notebooks"]
+exclude = ["worktrees", ".venv", "notebooks", "tests", "benchmark"]
 
 [[tool.mypy.overrides]]
-module = ["dask.*", "numba.*", "ducc0", "ducc0.*"]
+module = ["dask.*", "numba.*", "ducc0", "ducc0.*", "testing_utils", "matplotlib.*"]
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
