+++ tests/unit/hodges/test_detector.py
