+++ tests/unit/preprocessing/test_spectral.py
