+++ tests/unit/simple/test_detector.py
